mindcare-ai/backend/journal/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
transformers==4.41.2
deepface==0.0.89
openai-whisper==20231117
opencv-python==4.10.0.82
numpy==1.26.4